refactor(edges): clean up marker-alignment internals
- move test-only tangent readers to a testHelpers module
- rebuild plain bezier when the lead splice finds no cubic, so the
  fallback path still touches the real handles
- correct ARROW_LEAD_PX comment: marker lengths vary, 12 approximates
